Added deserialize benchmarks
