+++ src/GeoJSON.Text.Test.Benchmark/Deserialize/DeserializeFeatureLinestring.json
@@ -1,0 +1,25 @@
+﻿{
+  "type": "Feature",
+  "geometry": {
+    "coordinates": [
+      [
+        4.8892593383789062,
+        52.370725881211314
+      ],
+      [
+        4.8952674865722656,
+        52.3711451105601
+      ],
+      [
+        4.8920917510986328,
+        52.369310952782627
+      ],
+      [
+        4.8892593383789062,
+        52.370725881211314
+      ]
+    ],
+    "type": "LineString"
+  },
+  "properties": {}
+}
