+++ src/GeoJSON.Text.Test.Benchmark/Program.cs
@@ -1,6 +1,7 @@
 ﻿
 using BenchmarkDotNet.Running;
-using GeoJSON.Text.Test.Benchmark;
+using GeoJSON.Text.Test.Benchmark.Deserialize;
+using GeoJSON.Text.Test.Benchmark.Serialize;
 
 namespace GeoJSON.Text.Test.Benchmark
 {
@@ -9,8 +10,12 @@ namespace GeoJSON.Text.Test.Benchmark
         public static void Main(string[] args)
         {
             // Serialize
-            var summary1 = BenchmarkRunner.Run<BenchmarkSerializeFeatureCollectionLinestring>();
-            var summary2 = BenchmarkRunner.Run<BenchmarkSerializeFeatureLinestring>();
+            var summary1 = BenchmarkRunner.Run<SerializeFeatureCollectionLinestring>();
+            var summary2 = BenchmarkRunner.Run<SerializeFeatureLinestring>();
+
+            // Deserialize
+            var summary3 = BenchmarkRunner.Run<DeserializeFeatureLinestring>();
+            var summary4 = BenchmarkRunner.Run<DeserializeFeatureCollectionLinestring>();
         }
     }
 }
